requirements.txt
# Temel
numpy
pandas
matplotlib
tqdm
scikit-learn
pyyaml

# Görüntü işleme
opencv-python-headless
albumentations

# PyTorch ekosistemi
torch
torchvision
segmentation-models-pytorch

# Logging (opsiyonel)
tensorboard
wandb


jupyter